+++ source/engine/resources/ResourceManager.hpp
@@ -60,7 +60,7 @@ namespace engine {
 
         vk::DescriptorSetLayout getMeshDescriptorSetLayout();
 
-        uint64_t loadAnimation(const std::string& uri, const std::string& name);
+        uint32_t loadAnimation(const std::string& uri, const std::string& name);
 
     private:
         void createDescriptorPool();
@@ -73,7 +73,7 @@ namespace engine {
         std::unordered_map<uint64_t, engine::Texture> m_textures;
         std::unordered_map<uint64_t, std::shared_ptr<engine::Model>> m_models;
         std::unordered_map<uint64_t, engine::Mesh> m_meshes;
-        std::unordered_map<uint64_t, std::shared_ptr<Animation>> m_animations;
+        std::unordered_map<uint32_t, std::shared_ptr<Animation>> m_animations;
         std::vector<std::shared_ptr<engine::Shader>> m_shaders;
         vk::DescriptorPool m_imagesDescriptorPool{};
         vk::DescriptorSetLayout m_imagesDescriptorSetLayout{};
